Moves common constants to explore ones
