--- constants/common.js
@@ -1,14 +1,0 @@
-
-// Nav
-const EXPLORE_SECTIONS = [
-  { label: 'Solutions', query: { category: 'solutions' } },
-  { label: 'Investment capital', query: { category: 'investment-capital' } },
-  { label: 'Funding Sources', query: { category: 'funding-sources' } },
-  { label: 'Legal framework', query: { category: 'legal-framework' } },
-  { label: 'Investment components', query: { category: 'investment-components' } },
-  { label: 'Cities', query: { category: 'cities' } }
-];
-
-const EXPLORE_ROUTES = ['explore-index', 'explore-list', 'explore-detail'];
-
-export { EXPLORE_SECTIONS, EXPLORE_ROUTES };
