+++ constants/explore.js
@@ -1,4 +1,17 @@
 
+
+// Nav
+const EXPLORE_SECTIONS = [
+  { label: 'Solutions', query: { category: 'solutions' } },
+  { label: 'Investment capital', query: { category: 'investment-capital' } },
+  { label: 'Funding Sources', query: { category: 'funding-sources' } },
+  { label: 'Legal framework', query: { category: 'legal-framework' } },
+  { label: 'Investment components', query: { category: 'investment-components' } },
+  { label: 'Cities', query: { category: 'cities' } }
+];
+const EXPLORE_ROUTES = ['explore-index', 'explore-list', 'explore-detail'];
+
+
 // Tabs
 const EXPLORE_TABS = [
   { label: 'Solutions',
@@ -47,4 +60,4 @@ const EXPLORE_TABS = [
 ];
 
 
-export { EXPLORE_TABS };
+export { EXPLORE_ROUTES, EXPLORE_SECTIONS, EXPLORE_TABS };
